+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 -   Added support for __len__ for .NET classes that implement ICollection
 -   Added `object.GetRawPythonProxy() -> PyObject` extension method, that bypasses any conversions
 -   Added PythonException.Format method to format exceptions the same as traceback.format_exception
+-   Added Python 3 buffer api support and PyBuffer interface for fast byte and numpy array read/write ([#980][p980])
 
 ### Changed